requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
walletpay==1.3.1